Add and remove ingredients by checking and unchecking boxes
