+++ public/js/app.js
@@ -1,3 +1,7 @@
+require("dotenv").config();
+var unirest = require("unirest");
+
+// Ingredient category elements
 var dairyPage = $(".dairy");
 var meatPage = $(".meat");
 var veggiesPage = $(".veggies");
@@ -8,6 +12,159 @@ var recipesPage = $(".recipe-content");
 var submit = $(".submit");
 var heart = $(".heart");
 
+// Log in objects
+var localUser = {};
+var loginCheck = {};
+
+// The API object contains methods for each kind of request we'll make
+var API = {
+    // User calls
+    createUser: function(userObj) {
+      return $.ajax({
+        type: "POST",
+        url: "/api/user",
+        data: userObj,
+        success: function(response) {
+          localStorage.clear();
+          localStorage.setItem("userId", response.id);
+          localUser.id = localStorage.getItem("userId");
+          console.log(localUser.id);
+        }
+      });
+    },
+    getUser: function(userName) {
+      return $.ajax({
+        url: "/api/user/" + userName,
+        type: "GET",
+        success: function(res) {
+          loginCheck.id = res.id;
+          loginCheck.password = res.password;
+          console.log(res);
+        }
+      });
+    },
+    updateUser: function(userObj) {
+      return $.ajax({
+        url: "/api/user/" + userObj.id,
+        type: "PUT",
+        data: userObj
+      });
+    },
+
+    // fridge calls
+    getFridge: function(userId) {
+      return $.ajax({
+        url: "/api/fridge/" + userId,
+        type: "GET",
+        success: function(response) {
+          // check the necessary boxes
+          console.log(response);
+        }
+      });
+    },
+    addToFridge: function(fridgeObj) {
+      return $.ajax({
+        url: "/api/fridge",
+        type: "POST",
+        data: fridgeObj,
+        success: function(response) {
+          // Make the box checked
+          console.log(response);
+        }
+      });
+    },
+    deleteFromFridge: function(fridgeId) {
+      return $.ajax({
+        url: "/api/fridge/",
+        type: "DELETE",
+        data: fridgeId,
+        success: function(response) {
+          // Uncheck the box?
+          console.log(response);
+        }
+      });
+    },
+
+    // favorites calls
+    getFavorites: function(userId) {
+      return $.ajax({
+        url: "/api/favorites/" + userId,
+        type: "GET",
+        success: function(response) {
+          // create array of favorites
+          console.log(response);
+        }
+      });
+    },
+    addToFavorites: function(favoritesObj) {
+      return $.ajax({
+        url: "/api/favorites",
+        type: "POST",
+        data: favoritesObj,
+        success: function(response) {
+          // Check heart
+          console.log(response);
+        }
+      });
+    },
+    deleteFromFavorites: function(favoritesId) {
+        return $.ajax({
+            url: "/api/fridge/" + favoritesId,
+            type: "DELETE",
+            success: function(response) {
+                // Uncheck the box?
+                console.log(response);
+            }
+        });
+    },
+
+    // Spoonacular calls
+    searchForIngredient: function(text) {
+        unirest
+            .get(
+                "https://spoonacular-recipe-food-nutrition-v1.p.rapidapi.com/food/ingredients/autocomplete?number=10&intolerances=egg&query=" + text
+            )
+            .header(
+                "X-RapidAPI-Host",
+                "spoonacular-recipe-food-nutrition-v1.p.rapidapi.com"
+            )
+            .header("X-RapidAPI-Key", process.env.SPOONKEY)
+            .end(function(result) {
+                console.log(result.status, result.headers, result.body);
+            });
+    },
+    searchRecipeByIng: function(text) {
+        unirest
+            .get(
+                "https://spoonacular-recipe-food-nutrition-v1.p.rapidapi.com/recipes/findByIngredients?number=5&ranking=1&ignorePantry=false&ingredients=" + text
+            )
+            .header(
+                "X-RapidAPI-Host",
+                "spoonacular-recipe-food-nutrition-v1.p.rapidapi.com"
+            )
+            .header("X-RapidAPI-Key", process.env.SPOONKEY)
+            .end(function(result) {
+                console.log(result.status, result.headers, result.body);
+            });
+    },
+    searchRecipeById: function(id) {
+        unirest
+            .get(
+                "https://spoonacular-recipe-food-nutrition-v1.p.rapidapi.com/recipes/" + id + "/information"
+            )
+            .header(
+                "X-RapidAPI-Host",
+                "spoonacular-recipe-food-nutrition-v1.p.rapidapi.com"
+            )
+            .header("X-RapidAPI-Key", process.env.SPOONKEY)
+            .end(function(result) {
+                console.log(result.status, result.headers, result.body);
+            });
+    }
+
+};
+
+// Hide menues until selected
 function homePage() {
     dairyPage.hide();
     meatPage.hide();
@@ -16,12 +173,12 @@ function homePage() {
     seafoodPage.hide();
     fruitPage.hide();
     recipesPage.hide();
-    submit.hide();
-    
+    submit.hide(); 
 }
 
 homePage();
 
+// The following expand various menus depending on what the user selects
 function dairy() {
     var dairyClicked = $(".dairyBtn");
 
@@ -124,27 +281,26 @@ function fruits() {
 }
 fruits();
 
-function findRecipes() {
-    var submitClicked = $("#submit");
-
-    submitClicked.on('click', function () {
-        recipesPage.show();
-        heart.hide();
-    })
+// Set up temp user when user arrives to page
+function setTemp() {
+    if (localStorage.getItem("userId") === null) {
+        // Create temp user
+        var tempUser = {
+          name: "temp",
+          password: ""
+        };
+        localUser.name = tempUser.name;
+        console.log(tempUser);
+        // User post
+        API.createUser(tempUser);
+    } else {
+        localUser.id = localStorage.getItem("userId");
+        API.getFridge(localUser.id);
+    }
 }
+setTemp();
 
-findRecipes();
-
-function login() {
-    var loginClicked = $(".login");
-
-    loginClicked.on('click', function () {
-        $(".modal1").modal();
-    })
-}
-
-login();
-
+// Function that activates the modal for singing up
 function signup() {
     var signupClicked = $(".signup");
 
@@ -152,8 +308,93 @@ function signup() {
         $(".modal2").modal();
     })
 }
-
 signup();
+
+// Function for submitting sign up
+function signUpSubmit () {
+    $("#signup_submit").on("click", function(event) {
+        event.preventDefault();
+        console.log("button pressed");
+        console.log(
+        $("#user_name_signup")
+            .val()
+            .trim()
+        );
+        console.log($("#password_signup").val());
+        // create user object
+        var signUp = {
+        name: $("#user_name_signup")
+            .val()
+            .trim(),
+        password: $("#password_signup")
+            .val()
+            .trim()
+        };
+        var confirm = $("#password_confirm")
+        .val()
+        .trim();
+
+        console.log(signUp.name + " " + signUp.password + " " + confirm);
+        // check to make sure both inputs are present
+        if (!(signUp.name && signUp.password)) {
+        alert("You must enter an example text and description!");
+        return;
+        } else if (signUp.password !== confirm) {
+            alert("Password did not match");
+        } else {
+            // Create permanent user
+            if (localUser.name === "temp") {
+                signUp.id = localUser.id;
+                API.updateUser(signUp);
+            } else {
+                API.createUser(signUp);
+            }
+        }
+        // What happens when you submit sign up
+        // Right now, you are able to sign up with the same user name as someone else
+
+    });
+};
+signUpSubmit();
+
+// Function that activates the modal for logging in
+function login() {
+    var loginClicked = $(".login");
+
+    loginClicked.on('click', function () {
+        $(".modal1").modal();
+    })
+}
+login();
+
+// Function for submitting login
+function loginSubmit () {
+    // What happens when you submit log in
+    $("#login_submit").on("click", function(event) {
+        event.preventDefault();
+        // Create object from user input (name password)
+        var login = {
+            name: $("#user_name_login")
+                .val()
+                .trim(),
+            password: $("#password_login")
+                .val()
+                .trim()
+        };
+        // User GET with that object based on object.name
+        API.getUser(login.name);
+        if (login.password === loginCheck.password) {
+            localStorage.setItem("userId", login.id);
+            alert("Welcome, " + login.name + "!");
+        } else {
+            alert("Username or password did not match.");
+        }
+        // GET fridge based on id
+        // Set all the ingredients to their fridge values
+    });
+};
+loginSubmit();
+
 
 function clickSearch() {
     var searchClicked = $(".submit");
@@ -163,7 +404,16 @@ function clickSearch() {
     })
 }
 
+function findRecipes() {
+    var submitClicked = $("#submit");
+
+    submitClicked.on('click', function () {
+        recipesPage.show();
+        heart.hide();
+    })
+}
 clickSearch();
+findRecipes();
 
 $('#modal-login').click((event) => {
     $('#search').modal();
@@ -204,12 +454,26 @@ $("body").on("click", '.addition', function(event){
     $(`#${category}-ingredient`).val('');
     $(`#${category}-ingredient`).focus();
 
-
     console.log(input);
 })
 
+// When an ingredient item checkbox is clicked
 $('body').on('click', 'input[type="checkbox"]', function() {
-    console.log($(this).next().text());
+    var fridgeItem = {
+        name: $(this).next().text().toLowerCase(),
+        UserId: localStorage.getItem("userId")
+    }
+    console.log(fridgeItem);
+    // if they checked it, add the item
+    if ( $(this).prop("checked") ) {
+        console.log("Adding item");
+        API.addToFridge(fridgeItem);
+    } else {
+        // if they unchecked it, destroy the item
+        console.log("Deleting item");
+        console.log(fridgeItem);
+        API.deleteFromFridge(fridgeItem);
+    }
 });
 
 function favorite(){
@@ -220,5 +484,4 @@ function favorite(){
         heart.show();
     })
 }
-
 favorite();
